Refactor imports to use correct types
